+++ ex02/Class/Serializer/Serializer.hpp
@@ -5,15 +5,14 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: joakoeni <joakoeni@student.42mulhouse.fr>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/14 08:29:56 by joakoeni          #+#    #+#             */
-/*   Updated: 2024/03/14 08:30:04 by joakoeni         ###   ########.fr       */
+/*   Created: 2024/03/14 08:28:27 by joakoeni          #+#    #+#             */
+/*   Updated: 2024/04/11 15:04:50 by joakoeni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef SERIALIZER_HPP
 #define SERIALIZER_HPP
 
-#include <iostream>
 #include "../../Data.h"
 
 class Serializer
@@ -24,8 +23,8 @@ class Serializer
 		~Serializer();
 	public:
     Serializer& operator=(const Serializer& src);
-    static uintptr_t serialize(Data *ptr);
-    static Data *deserialize(uintptr_t raw);
+    static unsigned long serialize(Data *ptr); // Utilisation de unsigned long
+    static Data *deserialize(unsigned long raw); // Utilisation de unsigned long
 };
 
 #endif
